+++ packages/unlock-js/hardhat.config.js
@@ -4,10 +4,12 @@
 require('@nomiclabs/hardhat-ethers')
 require('@openzeppelin/hardhat-upgrades')
 
+const fs = require('fs-extra')
 const { subtask, task } = require('hardhat/config')
 const { HARDHAT_NETWORK_NAME } = require('hardhat/plugins')
 const { runCLI } = require('jest')
 const process = require('process')
+const path = require('path')
 const jestConfig = require('./jest.config')
 
 const TASK_JEST = 'test:jest'
@@ -25,13 +27,19 @@ subtask(TASK_JEST_RUN_TESTS).setAction(async () => {
 
 task(TASK_JEST, 'Runs jest tests').setAction(
   async ({ watch }, { run, network }) => {
-    // eslint-disable-next-line import/no-dynamic-require, global-require
-    const preCompile = require('./scripts/preCompile')
-
-    // pre-compile latest Unlock contract to be used by OZ upgrades
-    await preCompile({
-      unlockName: 'UnlockV10',
+    // copy Unlock contracts to be used by OZ upgrades
+    ;['UnlockV10', 'UnlockV11'].forEach(async (unlockName) => {
+      // copy contract source over
+      await fs.copy(
+        require.resolve(
+          `@unlock-protocol/contracts/dist/Unlock/${unlockName}.sol`
+        ),
+        path.resolve(`./src/__tests__/contracts/${unlockName}.sol`)
+      )
     })
+
+    // pre-compile contracts
+    await run('compile')
 
     const testResults = await run(TASK_JEST_RUN_TESTS, { watch })