+++ packages/unlock-js/scripts/compressAbi.js
@@ -13,7 +13,7 @@ const path = require('path')
 /* eslint-disable import/no-extraneous-dependencies */
 const abis = require('@unlock-protocol/contracts')
 
-const unlockVersions = ['v4', 'v6', 'v7', 'v8', 'v9', 'v10']
+const unlockVersions = ['v4', 'v6', 'v7', 'v8', 'v9', 'v10', 'v11']
 const publicLockVersions = ['v4', 'v6', 'v7', 'v8', 'v9']
 
 const data = {