+++ packages/unlock-js/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Non expired keys are represented as -1
 - Gas price calculation fixes
+- Adding support for unlimited durations when granting keys
 
 # 0.22.2